+++ check.py
@@ -45,235 +45,6 @@ from cjwmodule import i18n
 # * Column name ending in '_by': username
 # * Column name ending in '_at': timestamp
 
-SUBMISSIONS_AND_CLAIMS_SQL = r"""
-WITH
-smooch_requests AS (
-  SELECT
-    json_extract(daf.value_json, '$.source.type')
-      || ':'
-      || json_extract(daf.value_json, '$.source.originalMessageId')
-      AS submission_id,
-    json_extract(daf.value_json, '$.authorId') AS submitted_by,
-    a.created_at AS submitted_at,
-    json_extract(daf.value_json, '$.text') AS request_text,
-    a.annotated_id AS project_media_id
-  FROM dynamic_annotation_fields daf
-  INNER JOIN annotations a ON a.id = daf.annotation_id
-  WHERE daf.field_name = 'smooch_data'
-),
-web_requests AS (
-  SELECT
-    'check:' || project_medias.id AS submission_id,
-    'check:' || users.login as submitted_by,
-    project_medias.created_at AS submitted_at,
-    NULL AS request_text,
-    project_medias.id AS project_media_id
-  FROM project_medias
-  LEFT JOIN users ON project_medias.user_id = users.id
-  WHERE project_medias.id NOT IN (SELECT DISTINCT project_media_id FROM smooch_requests)
-),
-all_requests AS (
-  SELECT * FROM smooch_requests
-  UNION
-  SELECT * FROM web_requests
-),
-statuses AS (
-  SELECT
-    ROW_NUMBER() OVER (PARTITION BY a.annotated_id ORDER BY a.id DESC) AS rn_desc,
-    a.annotated_id AS project_media_id,
-    u.login AS login,
-    daf.value -- JSON-encoded String value
-  FROM dynamic_annotation_fields daf
-  INNER JOIN annotations a ON a.id = daf.annotation_id
-  LEFT JOIN users u ON a.annotator_id = u.id
-  WHERE a.annotation_type = 'verification_status'
-    AND a.annotated_type = 'ProjectMedia'
-    AND daf.field_name = 'verification_status_status'
-),
-last_statuses AS (
-  SELECT
-    project_media_id,
-    login,
-    json_extract(value, '$') AS status -- it's a JSON-encoded String
-  FROM statuses
-  WHERE rn_desc = 1
-),
-project_media_list1s AS (
-  SELECT
-    project_media_projects.project_media_id,
-    MIN(projects.title) AS list1
-  FROM project_media_projects
-  INNER JOIN projects ON project_media_projects.project_id = projects.id
-  GROUP BY project_media_projects.project_media_id
-),
-parent_relationships AS (
-  SELECT
-    relationships.target_id AS child_project_media_id,
-    relationships.source_id AS parent_project_media_id,
-    relationships.created_at,
-    users.login
-  FROM relationships
-  LEFT JOIN users ON relationships.user_id = users.id
-  WHERE relationships.relationship_type = '---' || CHAR(0xa) || ':source: parent' || CHAR(0xa) || ':target: child' || CHAR(0xa)
-),
-archived_events AS (
-  SELECT
-    ROW_NUMBER() OVER(PARTITION BY associated_id ORDER BY id ASC) AS rn,
-    associated_id AS project_media_id,
-    created_at,
-    whodunnit AS user_id
-  FROM versions
-  WHERE event_type = 'update_projectmedia'
-    AND associated_type = 'ProjectMedia'
-    AND json_extract(object_changes, '$.archived') = '[false,true]'
-),
-first_archived_events AS (
-  SELECT
-    archived_events.project_media_id,
-    archived_events.created_at,
-    users.login
-  FROM archived_events
-  LEFT JOIN users ON archived_events.user_id = users.id
-  WHERE archived_events.rn = 1
-),
-publish_related_events AS ( -- events that publish a report, edit a published report, or unpublish a report
-  SELECT
-    ROW_NUMBER() OVER(PARTITION BY associated_id ORDER BY id ASC) AS rn,
-    associated_id AS project_media_id,
-    created_at,
-    whodunnit AS user_id
-  FROM versions
-  WHERE event_type IN ('create_dynamic', 'update_dynamic')
-    AND associated_type = 'ProjectMedia'
-    AND object_changes LIKE '%\nstate: published\n%' -- JSON-encoded YAML value
-),
-first_publish_events AS (
-  SELECT
-    publish_related_events.project_media_id,
-    publish_related_events.created_at,
-    users.login
-  FROM publish_related_events
-  LEFT JOIN users ON publish_related_events.user_id = users.id
-  WHERE publish_related_events.rn = 1
-),
-status_change_events AS (
-  SELECT
-    ROW_NUMBER() OVER(PARTITION BY associated_id ORDER BY id ASC) AS rn,
-    ROW_NUMBER() OVER(PARTITION BY associated_id ORDER BY id DESC) AS rn_desc,
-    associated_id AS project_media_id,
-    created_at,
-    whodunnit AS user_id
-  FROM versions
-  WHERE event_type = 'update_dynamicannotationfield'
-    AND item_type = 'DynamicAnnotation::Field'
-    AND EXISTS (
-        SELECT 1
-        FROM dynamic_annotation_fields daf
-        WHERE daf.id = versions.item_id
-          AND daf.annotation_type = 'verification_status'
-          AND daf.field_name = 'verification_status_status'
-    )
-),
-first_status_change_events AS (
-  SELECT
-    status_change_events.project_media_id,
-    status_change_events.created_at,
-    users.login
-  FROM status_change_events
-  LEFT JOIN users ON status_change_events.user_id = users.id
-  WHERE status_change_events.rn = 1
-),
-last_status_change_events AS (
-  SELECT
-    status_change_events.project_media_id,
-    status_change_events.created_at,
-    users.login
-  FROM status_change_events
-  LEFT JOIN users ON status_change_events.user_id = users.id
-  WHERE status_change_events.rn_desc = 1
-),
-media_metadatas AS (
-  SELECT
-    annotations.annotated_id AS media_id,
-    dynamic_annotation_fields.value_json AS metadata_json
-  FROM dynamic_annotation_fields
-  INNER JOIN annotations ON dynamic_annotation_fields.annotation_id = annotations.id
-  WHERE annotations.annotated_type = 'Media'
-    AND annotations.annotation_type = 'metadata'
-    AND dynamic_annotation_fields.field_name = 'metadata_value'
-),
-project_media_metadatas AS (
-  SELECT
-    annotations.annotated_id AS project_media_id,
-    dynamic_annotation_fields.value_json AS metadata_json
-  FROM dynamic_annotation_fields
-  INNER JOIN annotations ON dynamic_annotation_fields.annotation_id = annotations.id
-  WHERE annotations.annotated_type = 'ProjectMedia'
-    AND annotations.annotation_type = 'metadata'
-    AND dynamic_annotation_fields.field_name = 'metadata_value'
-),
-reports AS (
-  SELECT
-    ROW_NUMBER() OVER(PARTITION BY annotated_id ORDER BY id DESC) AS rn_desc,
-    annotated_id AS project_media_id,
-    CASE WHEN annotations.data LIKE ('%' || CHAR(0xa) || 'state: published' || CHAR(0xa) || '%') THEN 'published' else 'paused' END AS status
-  FROM annotations
-  WHERE annotated_type = 'ProjectMedia'
-    AND annotation_type = 'report_design'
-),
-last_reports AS (
-  SELECT
-    project_media_id,
-    status
-  FROM reports
-  WHERE rn_desc = 1
-)
-SELECT
-  all_requests.submission_id AS "submission_id [text]",
-  all_requests.submitted_by,
-  all_requests.submitted_at,
-  all_requests.request_text,
-  project_medias.id AS claim_id,
-  last_statuses.status AS claim_status,
-  CASE last_statuses.login WHEN 'smooch' THEN NULL ELSE last_statuses.login END AS claim_status_by,
-  project_media_list1s.list1 AS claim_list1,
-  COALESCE(
-    json_extract(project_media_metadatas.metadata_json, '$.title'),
-    json_extract(media_metadatas.metadata_json, '$.title')
-  ) AS claim_title,
-  CASE medias.type WHEN 'Claim' THEN 'Text' ELSE medias.type END AS claim_type,
-  medias.url AS claim_url,
-  'https://checkmedia.org/' || teams.slug || '/media/' || project_medias.id AS check_url,
-  parent_relationships.parent_project_media_id AS primary_claim_id,
-  parent_relationships.created_at AS primary_claim_linked_at,
-  parent_relationships.login AS primary_claim_linked_by,
-  first_status_change_events.created_at AS first_claim_status_changed_at,
-  first_status_change_events.login AS first_claim_status_changed_by,
-  last_status_change_events.created_at AS last_claim_status_changed_at,
-  last_status_change_events.login AS last_claim_status_changed_by,
-  last_reports.status AS claim_report_status,
-  first_publish_events.created_at AS claim_report_first_published_at,
-  first_publish_events.login AS claim_report_first_published_by,
-  project_medias.archived AS "claim_archived [integer]",
-  first_archived_events.created_at AS claim_first_archived_at,
-  first_archived_events.login AS claim_first_archived_by
-FROM all_requests
-INNER JOIN project_medias ON project_medias.id = all_requests.project_media_id
-INNER JOIN medias ON project_medias.media_id = medias.id
-INNER JOIN teams ON teams.id = project_medias.team_id
-LEFT JOIN project_media_list1s ON project_media_list1s.project_media_id = project_medias.id
-LEFT JOIN last_statuses ON last_statuses.project_media_id = project_medias.id
-LEFT JOIN parent_relationships ON parent_relationships.child_project_media_id = project_medias.id
-LEFT JOIN first_status_change_events ON first_status_change_events.project_media_id = project_medias.id
-LEFT JOIN last_status_change_events ON last_status_change_events.project_media_id = project_medias.id
-LEFT JOIN project_media_metadatas ON project_media_metadatas.project_media_id = project_medias.id
-LEFT JOIN media_metadatas ON media_metadatas.media_id = medias.id
-LEFT JOIN last_reports ON last_reports.project_media_id = project_medias.id
-LEFT JOIN first_publish_events ON first_publish_events.project_media_id = project_medias.id
-LEFT JOIN first_archived_events ON first_archived_events.project_media_id = project_medias.id
-ORDER BY all_requests.submitted_at DESC, project_medias.id
-"""
-
 ITEMS_SQL = r"""
 WITH
 statuses AS (
@@ -738,13 +509,6 @@ def _cursor_to_table(cursor: sqlite3.Cursor) -> pa.Table:
     )
 
 
-def _query_submissions_and_claims(db: sqlite3.Connection) -> pa.Table:
-    """Return a table; raise sqlite3.ProgrammingError if queries fail."""
-    with contextlib.closing(db.cursor()) as cursor:
-        cursor.execute(SUBMISSIONS_AND_CLAIMS_SQL)
-        return _cursor_to_table(cursor)
-
-
 def _query_items(db: sqlite3.Connection) -> pa.Table:
     """Return a table; raise sqlite3.ProgrammingError if queries fail."""
     with contextlib.closing(db.cursor()) as cursor:
@@ -1113,10 +877,8 @@ def query_database(db: sqlite3.Connection, query_slug: str) -> pa.Table:
         return _query_tasks(db)
     elif query_slug == "conversations":
         return _query_conversations(db)
-    elif query_slug == "items":
-        return _query_items(db)
     else:
-        return _query_submissions_and_claims(db)
+        return _query_items(db)
 
 
 class InvalidLz4File(Exception):
@@ -1177,8 +939,9 @@ def _migrate_params_v0_to_v1(params):
     Valid slugs in v1:
 
     * submissions_and_claims
-    * tasks [NOT FINALIZED]
+    * tasks
     * conversations
+    * items
     """
     return {
         **params,
@@ -1186,9 +949,19 @@ def _migrate_params_v0_to_v1(params):
     }
 
 
+def _migrate_params_v1_to_v2(params):
+    """v1: `submissions_and_claims` is obsolete; v2 fallback to `items`."""
+    query_slug = params["query_slug"]
+    if query_slug == "submissions_and_claims":
+        query_slug = "items"
+    return {**params, "query_slug": query_slug}
+
+
 def migrate_params(params):
     if "query_slug" not in params:
         params = _migrate_params_v0_to_v1(params)
+    if params["query_slug"] == "submissions_and_claims":
+        params = _migrate_params_v1_to_v2(params)
     return params